Fix add_entities/subtract_entities normalization with unit conversion
When converting energy units, all the conversions take place in
energy.ts, and the results can be directly added together. However, when
_getMemoizedState was adding the values of add_entities, it didn't
realize that the entities to add were already the correct values, and it
erroneously normalized them. Now, the same "this has already been
converted" logic used for the main entity state is also used for the
add_entity state.