fix: connection_entity_id and add_entities/subtract_entities (#199)
* Fix connection_entity_id

If a connection_entity_id was specified, it was not added to the list of
entities for which to retrieve the state. It seems that
connection_entity_id only happened to work if that entity_id was also
used elsewhere.

* Fix add_entities/subtract_entities normalization with unit conversion

When converting energy units, all the conversions take place in
energy.ts, and the results can be directly added together. However, when
_getMemoizedState was adding the values of add_entities, it didn't
realize that the entities to add were already the correct values, and it
erroneously normalized them. Now, the same "this has already been
converted" logic used for the main entity state is also used for the
add_entity state.

* PR comments